+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 * Add `jaeger configuration reload` command. It's equivalent to using `jaeger configuration load --no-clone DESIGNID` where `DESIGNID` is the currently loaded design.
 * If called without arguments, `disable` now outputs the list of currently disabled robots.
 
+### ✨ Improved
+
+* The list of manually disabled positioners is kept during reinitialisation.
+
 ### 🏷️ Changed
 
 * Default FVC centroid algorithm is now `zbplus`.